Give refang the same treatment so both halves are reachable without the CLI (#14)
